refactor : UserInput vo 변경
